server/requirements.txt
fastapi
uvicorn
websockets
mcp[cli]
httpx
python-dotenv
anthropic
pydantic
python-multipart
pytest
pytest-asyncio